Tests issue micrometer with elastic-search on short lived app
Was able to replicate the issue identified in issue 1882.
Had to run the app around 10 times from the command line to see it replicated.
Please verify that this sample is using the correct versions.   I've been known to mess that up.

Using the following docker-compose to start an elastic-search instance.
https://github.com/jonatan-ivanov/local-services/blob/main/elasticsearch/docker-compose.yml